+++ examples/psb/000g_multiturn_inj.py
@@ -41,44 +41,9 @@ part_for_injection.py = df["y'"].values  * 1e-3 * (1 + part_for_injection.delta)
 part_for_injection.weight = 10
 
 
-class ParticlesInjectionSample:
 
-    def __init__(self, particles_to_inject, line, element_name, num_particles_to_inject):
-        self.particles_to_inject = particles_to_inject.copy()
-        self.line = line
-        self.element_name = element_name
-        self.num_particles_to_inject = num_particles_to_inject
 
-    def track(self, particles):
-
-        if not isinstance(particles._context, xo.ContextCpu):
-            raise ValueError('This element only works with CPU context')
-
-        self.particles_to_inject.at_turn += 1
-
-        if not self.num_particles_to_inject:
-            return
-
-        assert self.element_name in self.line.element_names
-        assert self.line[self.element_name] is self
-        at_element = self.line.element_names.index(self.element_name)
-        s_element = self.line.tracker._tracker_data_base.element_s_locations[at_element]
-
-        # Get random particles to inject
-        idx_inject = np.random.default_rng().choice(len(self.particles_to_inject.x),
-                                size=self.num_particles_to_inject, replace=False)
-
-        mask_inject = np.zeros(len(self.particles_to_inject.x), dtype=bool)
-        mask_inject[idx_inject] = True
-
-        p_inj = self.particles_to_inject.filter(mask_inject)
-        p_inj.s = s_element
-        p_inj.at_element = at_element
-        p_inj.update_p0c_and_energy_deviations(p0c=self.line.particle_ref.p0c[0])
-
-        particles.add_particles(p_inj)
-
-p_injection = ParticlesInjectionSample(particles_to_inject=part_for_injection,
+p_injection = xt.ParticlesInjectionSample(particles_to_inject=part_for_injection,
                                         line=line,
                                         element_name='injection',
                                         num_particles_to_inject=7)
